added db collection helper
